fix kill bug

+++ tools/bangu.py
@@ -98,9 +98,9 @@ elif args.opts == 'start':
 elif args.opts == 'stop':
     res = os.popen('ps -ef|grep bangu').readlines()
     for item in res:
-        if 'stop' not in item and 'grep bangu' not in item:
+        if 'stop' not in item and 'grep bangu' not in item and 'python' in item:
             pid = item.split()[1]
-            print item.replace('\n', ''), ' ---------------Killed!'
+            print item.replace('\n', ''), '\t---------------\tKilled!'
             os.system('kill -9 %s'% pid)
